% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_edu}
\alias{acs_edu}
\title{ACS Education Data}
\format{
A tibble with 1456 rows and 7 columns.
}
\usage{
acs_edu
}
\description{
A subset of the `acs_raw` tibble containing variables related to education.
}
\keyword{datasets}
