% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_housing}
\alias{acs_housing}
\title{ACS Housing Data}
\format{
A tibble with 1311 rows and 6 columns.
}
\usage{
acs_housing
}
\description{
A subset of the `acs_raw` tibble containing variables related to housing.
}
\keyword{datasets}
