% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_raw}
\alias{acs_raw}
\title{American Community Survey 5 year estimates}
\format{
A tibble with 1478 rows and 59 columns.
}
\usage{
acs_raw
}
\description{
A tibble containing ACS 5 year estimates provided by the Boston Area Research Initiative (BARI) for the years 2013-2017 by the census tract level.
}
\keyword{datasets}
