% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airbnb_hosts}
\alias{airbnb_hosts}
\title{Airbnb Hosts}
\format{
A tibble containing 1335 rows and 9 columns.
\describe{
   \item{id}{The host ID. Used for joining.}
   \item{name}{The host's name.}
   \item{since_year}{The year that they became a host.}
   \item{since_month}{The month that they became a host.}
   \item{since_day}{The day they became a host.}
   \item{reponse_rate}{The percent response rate of a host as a character vector. Requires cleaning.}
   \item{accemptance_rate}{The percent acceptance rate of a host as a character vector. Requires cleaning.}
   \item{superhost}{A dummy variable that indicates whether or not a host is a superhost.}
   \item{n_listings}{The number of a listings a host has.}
}
}
\usage{
airbnb_hosts
}
\description{
A tibble containing host information for Airbnb listings within the city of Boston.
}
\keyword{datasets}
