% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airbnb_listings}
\alias{airbnb_listings}
\title{Airbnb Listings}
\format{
A tibble with 3799 rows and 7 columns.
\describe{
   \item{id}{Listing ID. Used for joining.}
   \item{neighborhood}{The neighborhood the listing is located in.}
   \item{room_type}{The listing type—e.g. entire home, private room, etc.}
   \item{price}{The dollar price for a single night stay.}
   \item{minimum_nights}{The minimum number of nights a guest can stay at the listing.}
   \item{availability_365}{The number of days a year that the listing is available.}
   \item{host_id}{The ID of the host. Used for joining.}
}
}
\usage{
airbnb_listings
}
\description{
A tibble containing listings for Airbnb within the city of Boston.
}
\keyword{datasets}
