% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airbnb_locations}
\alias{airbnb_locations}
\title{Airbnb Locations}
\format{
A tibble with 3799 rows and 3 columns.
\describe{
   \item{id}{Listing ID. Used for joining.}
   \item{longitude}{The longitude of the listing in EPGS 4326.}
   \item{latitude}{The latitude of the listing in EPGS 4326.}
}
}
\usage{
airbnb_locations
}
\description{
A tibble containing the locations of Airbnb listings within the city of Boston.
}
\keyword{datasets}
