% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{commute}
\alias{commute}
\title{Greater Boston Area Commuting Data}
\format{
A tibble with 648 rows and 14 columns.
}
\usage{
commute
}
\description{
A subset of the `acs_raw` tibble containing commuting data for Middlesex, Norfolk, and Suffolk Counties.
}
\keyword{datasets}
