% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecometrics}
\alias{ecometrics}
\title{Ecometrics}
\format{
A tibble with 68 rows and 4 columns.
}
\usage{
ecometrics
}
\description{
A tibble containing frequency counts for BARI's annual release of Ecometrics for the years 2015 through 2018. See [here](https://ui.josiahparry.com/ecometrics.html) for more on ecometrics.
}
\keyword{datasets}
