% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suffolk_county}
\alias{suffolk_county}
\title{Suffolk County.}
\format{
An sf object with 203 rows and 1 variable `fips` in EPSG:4326.
 \describe{
   \item{fips}{Census tract fips code}
   \item{geometry}{Census tract geometry}
 }
}
\usage{
suffolk_county
}
\description{
An sf MULTIPOLYGON object representing the census tracts in Suffolk County, MA.
}
\keyword{datasets}
